Remove street stuff and rename MovementComponent

+++ src/systems/environmentSystem.cpp
@@ -34,9 +34,9 @@ void EnvironmentSystem::init() {
     for (int i = 0; i < 100; i++) {
         entt::entity entity = registry.create();
 
-        glm::vec3 position = glm::vec3((float)rand() / RAND_MAX * Configuration::worldSize, 0.0f, (float)rand() / RAND_MAX * Configuration::worldSize);
-        float angle = (float)rand() / RAND_MAX * 0.5f * M_PI;
-        glm::vec3 scale = glm::vec3((float)rand() / RAND_MAX * 0.5 + 1.5f);
+        glm::vec3 position = glm::vec3((float)rand() / static_cast<float>(RAND_MAX) * Configuration::worldSize, 0.0f, (float)rand() / static_cast<float>(RAND_MAX) * Configuration::worldSize);
+        float angle = (float)rand() / static_cast<float>(RAND_MAX) * 0.5f * M_PI;
+        glm::vec3 scale = glm::vec3((float)rand() / static_cast<float>(RAND_MAX) * 0.5 + 1.5f);
 
         registry.emplace<MeshComponent>(entity, treeGeometry, meshShader, treeTexture);
         registry.emplace<TransformationComponent>(entity,
